quasi finito il CommandListener
manca l'abbandona, il salva e il carica

+++ src/model/Lago.java
@@ -8,7 +8,7 @@ import javax.imageio.ImageIO;
 
 public class Lago implements Territorio {
 
-	private final static int BONUS=0; // bonus di permanenza delle unità sul territorio (DA RIDEFINIRE)
+	private final static double BONUS=0; // bonus di permanenza delle unità sul territorio (DA RIDEFINIRE)
 	private BufferedImage bImg; // immagine della lago
 	private final static int costo= Integer.MAX_VALUE; // non verrà mai attraversato
 	
@@ -30,7 +30,7 @@ public class Lago implements Territorio {
 		return costo;
 	}
 	
-	public int getBonus(){
+	public double getBonus(){
 		return BONUS;
 	}
