+++ src/model/Territorio.java
@@ -5,6 +5,6 @@ import java.awt.Image;
 public interface Territorio{
 	public String getNome();
 	public int getCosto();
-	public int getBonus();
+	public double getBonus();
 	public Image getImage();
 }
